requirements.txt
grpclib
protobuf
grpclib
grpcio-tools
uvloop
pyyaml
black
structlog
